Refactor /xldata window
