+++ Dalamud/Interface/Internal/Windows/Data/ConfigurationWidget.cs
@@ -1,0 +1,29 @@
+﻿using Dalamud.Configuration.Internal;
+using Dalamud.Utility;
+
+namespace Dalamud.Interface.Internal.Windows;
+
+/// <summary>
+/// Widget for displaying configuration info.
+/// </summary>
+internal class ConfigurationWidget : IDataWindowWidget
+{
+    /// <inheritdoc/>
+    public DataKind DataKind { get; init; } = DataKind.Configuration;
+
+    /// <inheritdoc/>
+    public bool Ready { get; set; }
+
+    /// <inheritdoc/>
+    public void Load()
+    {
+        this.Ready = true;
+    }
+
+    /// <inheritdoc/>
+    public void Draw()
+    {
+        var config = Service<DalamudConfiguration>.Get();
+        Util.ShowObject(config);
+    }
+}
