+++ Dalamud/Interface/Internal/Windows/Data/PartyListWidget.cs
@@ -1,0 +1,69 @@
+﻿using System;
+
+using Dalamud.Game.ClientState.Objects.SubKinds;
+using Dalamud.Game.ClientState.Objects.Types;
+using Dalamud.Game.ClientState.Party;
+using Dalamud.Utility;
+using ImGuiNET;
+
+namespace Dalamud.Interface.Internal.Windows;
+
+/// <summary>
+/// Widget for displaying information about the current party.
+/// </summary>
+internal class PartyListWidget : IDataWindowWidget
+{
+    private bool resolveObjects;
+    private bool resolveGameData;
+
+    /// <inheritdoc/>
+    public DataKind DataKind { get; init; } = DataKind.Party_List;
+
+    /// <inheritdoc/>
+    public bool Ready { get; set; }
+
+    /// <inheritdoc/>
+    public void Load()
+    {
+        this.Ready = true;
+    }
+
+    /// <inheritdoc/>
+    public void Draw()
+    {
+        var partyList = Service<PartyList>.Get();
+
+        ImGui.Checkbox("Resolve GameData", ref this.resolveGameData);
+        ImGui.Checkbox("Resolve Actors", ref this.resolveObjects);
+
+        ImGui.Text($"GroupManager: {partyList.GroupManagerAddress.ToInt64():X}");
+        ImGui.Text($"GroupList: {partyList.GroupListAddress.ToInt64():X}");
+        ImGui.Text($"AllianceList: {partyList.AllianceListAddress.ToInt64():X}");
+
+        ImGui.Text($"{partyList.Length} Members");
+
+        for (var i = 0; i < partyList.Length; i++)
+        {
+            var member = partyList[i];
+            if (member == null)
+            {
+                ImGui.Text($"[{i}] was null");
+                continue;
+            }
+
+            ImGui.Text($"[{i}] {member.Address.ToInt64():X} - {member.Name} - {member.GameObject?.ObjectId}");
+            if (this.resolveObjects)
+            {
+                var actor = member.GameObject;
+                if (actor == null)
+                {
+                    ImGui.Text("Actor was null");
+                }
+                else
+                {
+                    Util.PrintGameObject(actor, "-", this.resolveGameData);
+                }
+            }
+        }
+    }
+}
